SkillSight/requirements.txt
flask
flask-cors
gtts
mysql-connector-python
python-dotenv
gunicorn
